# Conflicts:
#	docs/docker/qp.md
#	docs/nuget/Quantumart.QP8.BLL.md

+++ siteMvc/Controllers/UploadController.cs
@@ -3,11 +3,12 @@ using System.IO;
 using System.Net;
 using System.Threading.Tasks;
 using Microsoft.AspNetCore.Http;
-using Microsoft.AspNetCore.Http.Features;
 using Microsoft.AspNetCore.Mvc;
+using Minio;
 using NLog;
 using NLog.Fluent;
 using Quantumart.QP8.BLL;
+using Quantumart.QP8.BLL.Helpers;
 using Quantumart.QP8.BLL.Repository;
 using Quantumart.QP8.Configuration;
 using Quantumart.QP8.Constants;
@@ -19,12 +20,32 @@ namespace Quantumart.QP8.WebMvc.Controllers
 {
     public class UploadController : AuthQpController
     {
-        private readonly IBackendActionLogRepository _logger;
+        private readonly IBackendActionLogRepository _logRepository;
         private static readonly ILogger Logger = LogManager.GetCurrentClassLogger();
+        private readonly S3Options _options;
+        private readonly PathHelper _pathHelper;
 
-        public UploadController(IBackendActionLogRepository logger)
+        public UploadController(
+            IBackendActionLogRepository logRepository,
+            S3Options options,
+            PathHelper pathHelper
+        )
         {
-            _logger = logger;
+            _logRepository = logRepository;
+            _options = options;
+            _pathHelper = pathHelper;
+            EnsureS3ClientCreated();
+        }
+
+        private void EnsureS3ClientCreated()
+        {
+            if (_pathHelper.UseS3)
+            {
+                new MinioClient()
+                    .WithEndpoint(_options.Endpoint)
+                    .WithCredentials(_options.AccessKey, _options.SecretKey)
+                    .Build();
+            }
         }
 
         private void LogError(string msg, string fileName, Exception ex = null)
@@ -55,21 +76,29 @@ namespace Quantumart.QP8.WebMvc.Controllers
                 throw new ArgumentException("Folder Path is empty");
             }
 
-            if (!Directory.Exists(destinationUrl))
+            if (_pathHelper.UseS3 && !destinationUrl.StartsWith(QPConfiguration.TempDirectory))
             {
-                Directory.CreateDirectory(destinationUrl);
+                destinationUrl = _pathHelper.FixPathSeparator(destinationUrl);
+            }
+            else
+            {
+                if (!Directory.Exists(destinationUrl))
+                {
+                    Directory.CreateDirectory(destinationUrl);
+                }
             }
 
-            chunk = chunk ?? 0;
-            chunks = chunks ?? 1;
+            chunk ??= 0;
+            chunks ??= 1;
             PathSecurityResult securityResult;
 
             var tempPath = Path.Combine(QPConfiguration.TempDirectory, name);
-            var destPath = Path.Combine(destinationUrl, name);
+            var destPath = _pathHelper.CombinePath(destinationUrl, name);
+
 
             if (chunk == 0 && chunks == 1)
             {
-                securityResult = PathInfo.CheckSecurity(destinationUrl);
+                securityResult = PathInfo.CheckSecurity(destinationUrl, true, _pathHelper.Separator);
                 if (!securityResult.Result)
                 {
                     var errorMsg = string.Format(PlUploadStrings.ServerError, name, destinationUrl, $"Access to the folder (ID = {securityResult.FolderId}) denied");
@@ -81,10 +110,18 @@ namespace Quantumart.QP8.WebMvc.Controllers
 
                 try
                 {
-                    using (var fileStream = new FileStream(destPath, FileMode.Create))
+                    if (_pathHelper.UseS3 && !destPath.StartsWith(QPConfiguration.TempDirectory))
                     {
+                        await using var stream = file.OpenReadStream();
+                        _pathHelper.SetS3File(stream, destPath);
+                    }
+                    else
+                    {
+                        await using var fileStream = new FileStream(destPath, FileMode.Create);
                         await file.CopyToAsync(fileStream);
                     }
+
+                    CreateLogs(name, destPath, securityResult);
                 }
                 catch (Exception ex)
                 {
@@ -99,10 +136,8 @@ namespace Quantumart.QP8.WebMvc.Controllers
             {
                 try
                 {
-                    using (var fileStream = new FileStream(tempPath, chunk == 0 ? FileMode.Create : FileMode.Append))
-                    {
-                       await file.CopyToAsync(fileStream);
-                    }
+                    await using var fileStream = new FileStream(tempPath, chunk == 0 ? FileMode.Create : FileMode.Append);
+                    await file.CopyToAsync(fileStream);
                 }
                 catch (Exception ex)
                 {
@@ -118,9 +153,7 @@ namespace Quantumart.QP8.WebMvc.Controllers
                     var isTheLastChunk = chunk.Value == chunks.Value - 1;
                     if (isTheLastChunk)
                     {
-                        securityResult = PathInfo.CheckSecurity(destinationUrl);
-                        var actionCode = securityResult.IsSite ? ActionCode.UploadSiteFile : ActionCode.UploadContentFile;
-
+                        securityResult = PathInfo.CheckSecurity(destinationUrl, true, _pathHelper.Separator);
                         if (!securityResult.Result)
                         {
                             var errorMsg = string.Format(PlUploadStrings.ServerError, name, destinationUrl, $"Access to the folder (ID = {securityResult.FolderId}) denied");
@@ -130,19 +163,23 @@ namespace Quantumart.QP8.WebMvc.Controllers
                             return Json(new { message = errorMsg, isError = true });
                         }
 
-                        if (FileIO.Exists(destPath))
+                        if (_pathHelper.UseS3)
                         {
-                            FileIO.SetAttributes(destPath, FileAttributes.Normal);
-                            FileIO.Delete(destPath);
+                            await using var tempStream = new FileStream(tempPath, FileMode.Open, FileAccess.Read);
+                            _pathHelper.SetS3File(tempStream, destPath);
+                        }
+                        else
+                        {
+                            if (FileIO.Exists(destPath))
+                            {
+                                FileIO.SetAttributes(destPath, FileAttributes.Normal);
+                                FileIO.Delete(destPath);
+                            }
+
+                            FileIO.Move(tempPath, destPath!);
                         }
 
-                        FileIO.Move(tempPath, destPath);
-                        BackendActionContext.SetCurrent(actionCode, new[] { name }, securityResult.FolderId);
-
-                        var logs = BackendActionLog.CreateLogs(BackendActionContext.Current, _logger);
-                        _logger.Save(logs);
-
-                        BackendActionContext.ResetCurrent();
+                        CreateLogs(name, destPath, securityResult);
                     }
                 }
                 catch (Exception ex)
@@ -158,6 +195,17 @@ namespace Quantumart.QP8.WebMvc.Controllers
             }
 
             return Json(new { message = $"file{name} uploaded", isError = false });
+        }
+
+        private void CreateLogs(string name, string path, PathSecurityResult securityResult)
+        {
+            if (!path.StartsWith(QPConfiguration.TempDirectory))
+            {
+                var actionCode = securityResult.IsSite ? ActionCode.UploadSiteFile : ActionCode.UploadContentFile;
+                BackendActionContext.CreateLogs(
+                    actionCode, new[] { name }, securityResult.FolderId, _logRepository, false
+                );
+            }
         }
     }
 }
